Adapt Dyn-HaMR for PyTorch 2.7, VIPE camera integration, and multi-chunk HMP fixes
- Patch torch.load for weights_only=True default (PyTorch 2.7)
- Add VIPE camera loading and video_vipe data config
- Fix HMP multi-chunk fitting (StepLR verbose removal, chunk boundary handling)
- Fix checkpoint resumption (re-set requires_grad after loading params)
- Fix dtype mismatch in output.py

+++ dyn-hamr/optim/output.py
@@ -173,7 +173,7 @@ def save_camera_json(path, cam_R, cam_t, intrins):
     :param intrins (N, 4)
     """
     N = len(cam_R)
-    T = torch.tensor([[1, 0, 0], [0, -1, 0], [0, 0, -1]], dtype=torch.float32)
+    T = torch.tensor([[1, 0, 0], [0, -1, 0], [0, 0, -1]], dtype=cam_R.dtype)
     cam_R = torch.einsum("ij,bjk->bik", T, cam_R)
     cam_t = torch.einsum("ij,bj->bi", T, cam_t)
     with open(path, "w") as f:
